hospital-readmission-fairness/pyproject.toml
[project]
name = "readmission-fairness"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "python-dotenv>=1.0",
  "typer>=0.9",
  "pydantic>=2.0",
  "pytest>=7.0"
]

[project.scripts]
readmission = "readmission.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
